# isanzure-backend/app/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Isanzure API",
    "description": "Isanzure backend REST API documentation",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "UserRole": {
        "type": "string",
        "enum": ["super_admin", "rental_company", "agency", "customer"],
        "default": "customer"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "tags": ["System"],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "example": { "status": "healthy", "environment": "development" }
              }
            }
          }
        }
      }
    },
    "/api/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "phone", "password"],
                "properties": {
                  "name":     { "type": "string", "example": "John Doe" },
                  "email":    { "type": "string", "example": "john@example.rw" },
                  "phone":    { "type": "string", "example": "+250788123456" },
                  "password": { "type": "string", "example": "securepassword123" },
                  "role":     { "$ref": "#/components/schemas/UserRole" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "example": { "user_id": "uuid", "role": "customer", "token": "jwt_token" }
              }
            }
          },
          "400": { "description": "Missing fields", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login and receive a JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email":    { "type": "string", "example": "john@example.rw" },
                  "password": { "type": "string", "example": "securepassword123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "example": { "token": "jwt_token", "role": "customer" }
              }
            }
          },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request a password reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "example": "john@example.rw" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset token generated",
            "content": {
              "application/json": {
                "example": { "message": "Password reset token generated.", "reset_token": "token" }
              }
            }
          },
          "400": { "description": "Missing fields", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Reset password using a reset token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["reset_token", "new_password"],
                "properties": {
                  "reset_token":  { "type": "string", "example": "token_from_forgot_password" },
                  "new_password": { "type": "string", "example": "newpassword123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "example": { "message": "Password reset successful." }
              }
            }
          },
          "400": { "description": "Invalid or expired token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Get current user profile",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "example": {
                  "id": "uuid",
                  "name": "John Doe",
                  "email": "john@example.rw",
                  "phone": "+250788123456",
                  "role": "customer",
                  "created_at": "2026-07-07T00:00:00"
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "User not found" }
        }
      },
      "put": {
        "tags": ["Profile"],
        "summary": "Update current user profile",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "phone"],
                "properties": {
                  "name":  { "type": "string", "example": "Jane Doe" },
                  "phone": { "type": "string", "example": "+250788654321" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "example": { "id": "uuid", "name": "Jane Doe", "email": "john@example.rw", "phone": "+250788654321", "role": "customer" }
              }
            }
          },
          "400": { "description": "Missing fields" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  }
}